+++ .gitignore
@@ -1,0 +1,3 @@
+1.1.2'
+/lib/lita/hanlders/.byebug_history
+/lib/lita/handlers/danger.rb
