--- README.md
@@ -1,70 +1,0 @@
-# RUBY CAPSTONE PROJECT
-
-In this project I wanted to build a corona and city bot, a bot that helps you to find informations about a country you asked
-
-## Built With
-
-- Ruby
-- lita
-- redis
-- nokogiri
-- httparty
-- VSCode, Linters, Rubocop
-
-
-## Getting Started
-
-To get a local copy up and running follow these simple example steps.
-
-### Prerequisites
-
-- Get ruby latest version installed
-
-### Setup
-
-- Clone the repository on your local machine
-- cd into the folder, find lita_corona_info folder search for the bin/main.rb file an open it in code editor
-- run it whith main.rb command or ruby main.rb command
-
-### Install
-
-- Install VSCode or any code editor you like
-- Install Ruby(most recent version)
-- Run this command on your terminal in order to install rubocop: gem install rubocop 
-
-### Usage
-
-### Run tests
-
-- you will find comment lines at the end of the project, just uncomment them
-- run the file using "ruby filename" command
-- install lita, nokogiri, httparty  whith gem install lita, nokogiri, httparty command
-- install redis sever
-
-## Authors
-
-
-👤 **Theophile Kango**
-
-- Github: [@githubhandle](https://github.com/Theophile-Kango)
-- Twitter: [@twitterhandle](https://twitter.com/Theophadh)
-- Linkedin: [linkedin](https://www.linkedin.com/in/theophile-kango-b6b580194/)
-
-## 🤝 Contributing
-
-Contributions, issues and feature requests are welcome!
-
-Feel free to check the [issues page](https://github.com/Theophile-Kango/ruby_capston_project/issues).
-
-## Show your support
-
-Give a ⭐️ if you like this project!
-
-## Acknowledgments
-
-- Hat tip to anyone whose code was used
-- Inspiration
-- etc
-
-## 📝 License
-
