+++ lib/lita/handlers/corona_info.rb
