--- lita_corona_info/Gemfile
@@ -1,3 +1,0 @@
-source "https://rubygems.org"
-
-gemspec
