--- lita_corona_info/README.md
@@ -1,19 +1,0 @@
-# lita-corona_info
-
-TODO: Add a description of the plugin.
-
-## Installation
-
-Add lita-corona_info to your Lita instance's Gemfile:
-
-``` ruby
-gem "lita-corona_info"
-```
-
-## Configuration
-
-TODO: Describe any configuration attributes the plugin exposes.
-
-## Usage
-
-TODO: Describe the plugin's features and how to use them.
