--- lita_corona_info/lib/lita/handlers/.byebug_history
@@ -1,256 +1,0 @@
-exit
-get_info("china")
-get_name("china")
-exit
-get_info("india")
-get_info("china")
-exit
-get_info("china")
-exit
-get_info("china")
-exit
-get_info("china")
-@sort_list.compact[0][0]["Country"]
-@sort_list.compact[0][0]
-@sort_list[0][0]["china"]
-@sort_list.compact[0][0]["china"]
-get_info("china")
-exit
-get_info("china")
-exit
-get_info("china")
-exit
-th2
-th2.text
-th2 = thead.css('th').text
-th2 = thead.css('th')
-th = thead.css('th').text
-th = thread.css('th').text
-th = thead.css('th')
-th = thead.css('th').text
-thead = table.css('thead')
-table = parsed_page.css('table#main_table_countries_today')
-parsed_page.css('table#main_table_countries_today').text
-parsed_page.css('table#main_table_countries_today')
-get_info("china")
-exit
-get_info("china")
-exit
-get_info("china")
-exit
-get_info("china")
-exit
-get_info("china")
-exit
-set_regex([1])
-set_regex("pp")
-get_reg
-exit
-set_regex(element)
-set_regex(@regex)
-get_reg
-exit
-get_reg
-exit
-get_regex
-exit
-@sort_list.compact[0][0]["Country"]
-@sort_list.compact[0][0]
-@sort_list.compact[0][0]["country"]
-@sort_list.compact[0]
-@sort_list[0]
-@sort_list
-@sort_list[0]
-@sort_list[0][0]["country"]
-@sort_list.compact[0]["country"]
-@sort_list.compact[0][0]["country"]
-get_reg
-exit
-get_reg
-@regex
-get_regex
-exit
-get_info("china")
-exit
-get_info("rwanda")
-get_info("congo")
-exit
-get_info("congo")
-exit
-get_info("france")
-get_info("rwanda")
-get_info("Congo")
-get_info("congo")
-get_info("china")
-exit
-get_info("china")
-exit
-get_info("china")
-exit
-get_info("congo")
-exit
-get_info("china")
-exit
-@sort_list.compact[0][0]["Country"]
-@sort_list.compact[0]
-@sort_list.compact
-@sort_list[0]
-@sort_list.class[0]
-@sort_list.class
-@sort_list
-@sort_list.first
-@sort_list.start
-@sort_list[0]
-@sort_list[0][0]["Country"].text
-@sort_list[0]
-@sort_list[0][0]["Country"]
-get_info("china")
-@sort_list
-details_array
-details.text
-details
-parsed_page.text
-parsed_page
-exit
-get_info("china")
-exit
-get_info("china")
-exit
-get_info("china")
-info("china")
-exit
-get_info("china")
-exit
-get_info('china')
-exit
-get_info("china")
-exit
-get_info("china")
-exit
-get_info("russia")
-exit
-get_info("kenya")
-exit
-get_info("Rwanda")
-exit
-get_info("congo")
-exit
-get_info("china")
-exit
-cls
-exit
-get_info("china")
-exit
-get_info("China")
-exit
-get_sort_list[0][0]["Country"]
-exit
-get_sort_list
-exit
-get_sort_list
-exit
-get_sort_list
-exit
-get_sort_list
-exit
-get_sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-cls
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list.length
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-details_array[0][0]
-exit
-sort_list
-exit
-cls
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-cls
-sort_list
-exit
-sort_list
-exit
-sort_list
-sort_array
-exit
-details_array
-details_arr
-exit
-sort_list
-exit
-sort_list
-exit
-details_array
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-sort_list
-exit
-details_array
-exit
-details_array
-exit
-details_array
-exit
-details_array
-exit
-details_array
-exit
-details_arr
-exit
-details_arr
-exit
-cls
-i
-details_arr
-details_array
-exit
-details_arr
-details_array
