--- lita_corona_info/locales/en.yml
@@ -1,4 +1,0 @@
-en:
-  lita:
-    handlers:
-      corona_info:
