--- lita_corona_info/spec/lita/handlers/corona_info_spec.rb
@@ -1,11 +1,0 @@
-# rubocop:disable all
-require_relative './spec_helper'
-
-describe Lita::Handlers::CoronaInfo, lita_handler: true do
-  context 'when we write hello' do
-    it { is_expected.to route('hello').to(:hellowrod) }
-    it { is_expected.not_to route('morning').to(:hellowrod) }
-  end
-end
-
-
