--- lita_corona_info/spec/spec_helper.rb
@@ -1,6 +1,0 @@
-require_relative "lita-corona_info"
-require_relative "lita/rspec"
-
-# A compatibility mode is provided for older plugins upgrading from Lita 3. Since this plugin
-# was generated with Lita 4, the compatibility mode should be left disabled.
-Lita.version_3_compatibility_mode = false
