+++ locales/en.yml
@@ -1,0 +1,4 @@
+en:
+  lita:
+    handlers:
+      corona_info:
